pyproject.toml
[project]
name = "youtube-timestamps"
version = "0.1.0"
description = "Generate chapter timestamps for YouTube videos using AI transcription"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Craig Mod", email = "craig@craigmod.com" }
]
keywords = ["youtube", "timestamps", "chapters", "whisper", "gpt", "transcription", "cli", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Utilities",
]
dependencies = [
    "yt-dlp>=2024.12.13",
    "openai>=1.57.0",
    "google-genai>=1.0.0",
    "pydub>=0.25.1",
    "audioop-lts>=0.2.1; python_version >= '3.13'",
    "click>=8.1.7",
    "rich>=13.9.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]

[project.urls]
Repository = "https://github.com/craigmod/youtube-timestamps"
"Bug Tracker" = "https://github.com/craigmod/youtube-timestamps/issues"

[project.scripts]
youtube-timestamps = "main:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
